requirements.txt
django==5.0.6
djangorestframework==3.15.1
pika==1.3.2
django-cors-headers==4.3.1
django-mysql==4.13.0
mysqlclient==2.2.4
pillow==10.3.0